requirements.txt
# Web Framework and Server
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0
streamlit>=1.0.0

# Data Processing and Visualization
pandas>=1.3.0
plotly>=5.3.0
numpy>=1.21.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0

# Computer Vision
opencv-python>=4.5.0

# PDF Generation
fpdf>=1.7.2
reportlab>=3.6.0

# Environment Variables
python-dotenv>=0.19.0
